test(ui): split chat page coverage
